Expose canvas core in brush and stroke
