+++ Assets/Scripts/AppCore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee6349b6a44b4b98a5d4712d84cbf9b5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
