+++ Assets/Scripts/Brushes/BaseBrush.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6392d62c-e01c-431e-ad94-7701f81ba65a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
