+++ Assets/Scripts/MinimalExampleCore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04e80df073b643cf8a213ce7b33d16ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
